Create tidy data set from data collected from the accelerometers from the Samsung Galaxy S smartphone
This script utilizes a data set collected from the accelerometers from
the Samsung Galaxy S smartphone by merging the training and the test
sets to create one data set. It includes only the measurements on the
mean and standard deviation for each measurement. It changes the
activity ID to descriptive activity names to name the activities in the
data set. It also labels the data set with descriptive variable names,
followed by creating a second, independent tidy data set with the
average of each variable for each activity and each subject. The final
result is saved in a file called “tidayData.txt”.